Completely decoupled autologin
