+++ include/mq/Proto.h
@@ -1,0 +1,24 @@
+/*
+ * MacroQuest: The extension platform for EverQuest
+ * Copyright (C) 2002-2020 MacroQuest Authors
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License, version 2, as published by
+ * the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ */
+
+#pragma once
+
+#include "../../src/common/Mailbox.h"
+#include "../../src/common/proto/Shared.pb.h"
+
+#ifdef _DEBUG
+#pragma comment(lib, "libprotobufd")
+#else
+#pragma comment(lib, "libprotobuf")
+#endif
