+++ src/main/PipeClient.h
@@ -25,13 +25,6 @@ void ShutdownPipeClient();
 
 namespace pipeclient {
 
-MQLIB_API void NotifyCharacterLoad(const char* Profile, const char* Account, const char* Server, const char* Character);
-MQLIB_API void NotifyCharacterUnload(const char* Profile, const char* Account, const char* Server, const char* Character);
-MQLIB_API void NotifyCharacterUpdate(int Class, int Level);
-MQLIB_API void LoginServer(const char* Login, const char* Pass, const char* Server);
-MQLIB_API void LoginCharacter(const char* Login, const char* Pass, const char* Server, const char* Character);
-MQLIB_API void LoginProfile(const char* Profile, const char* Server, const char* Character);
-
 MQLIB_API uint32_t GetLauncherProcessID();
 
 void NotifyIsForegroundWindow(bool isForeground);
